debug all the issues
