+++ frontend/src/routes/PublicRoutes.jsx
@@ -8,10 +8,12 @@ import TrackOrderForm from "../pages/TrackOrderForm";
 import AboutUs from "../pages/AboutUs";
 import { Support } from "@mui/icons-material";
 import SupportPage from "../pages/SupportPage";
+import ForgetPassword from "../components/logins/ForgetPassword";
 const PublicRoutes = (
   <>
     <Route path="/" element={<Home />} />
     <Route path="/login" element={<User />} />
+    <Route path="/forgot-password" element={<ForgetPassword/>} />
     <Route path="/signup" element={<Signup />} />
        <Route path="/track-order" element={<TrackOrderForm/>} />
        <Route path="/about-us" element={<AboutUs/>} />
